+++ src/main/java/com/ing/data/cassandra/jdbc/Utils.java
@@ -116,6 +116,10 @@ public final class Utils {
      */
     public static final String KEY_PASSWORD = "password";
     /**
+     * JDBC URL parameter key for the request timeout.
+     */
+    public static final String KEY_REQUEST_TIMEOUT = "requesttimeout";
+    /**
      * JDBC URL parameter key for the configuration file.
      */
     public static final String KEY_CONFIG_FILE = "configfile";
@@ -140,6 +144,7 @@ public final class Utils {
     public static final String TAG_SSL_ENGINE_FACTORY = "sslEngineFactory";
     public static final String TAG_CLOUD_SECURE_CONNECT_BUNDLE = "secureConnectBundle";
     public static final String TAG_CONFIG_FILE = "configFile";
+    public static final String TAG_REQUEST_TIMEOUT = "requestTimeout";
 
     public static final String JSSE_TRUSTSTORE_PROPERTY = "javax.net.ssl.trustStore";
     public static final String JSSE_TRUSTSTORE_PASSWORD_PROPERTY = "javax.net.ssl.trustStorePassword";
@@ -308,6 +313,9 @@ public final class Utils {
                 }
                 if (params.containsKey(KEY_PASSWORD)) {
                     props.setProperty(TAG_PASSWORD, params.get(KEY_PASSWORD));
+                }
+                if (params.containsKey(KEY_REQUEST_TIMEOUT)) {
+                    props.setProperty(TAG_REQUEST_TIMEOUT, params.get(KEY_REQUEST_TIMEOUT));
                 }
                 if (params.containsKey(KEY_CONFIG_FILE)) {
                     props.setProperty(TAG_CONFIG_FILE, params.get(KEY_CONFIG_FILE));